:fire: remove Role

+++ src/main/java/com/backend/chatop4/model/Rental.java
@@ -19,14 +19,14 @@ import lombok.Setter;
 public class Rental {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Integer id;
 
   private String name;
   private Integer price;
   private Integer surface;
   private String picture;
   private String description;
-  private Long owner_id;
+  private Integer owner_id;
   private Date created_at;
   private Date updated_at;
 }
